+++ dart/lib/src/default_integrations.dart
@@ -13,22 +13,59 @@ class RunZonedGuardedIntegration extends Integration {
 
   final Future<void> Function() _runner;
 
+  /// Needed to check if we somehow caused a `print()` recursion
+  bool _isPrinting = false;
+
   @override
   FutureOr<void> call(Hub hub, SentryOptions options) {
-    runZonedGuarded(() async {
-      await _runner();
-    }, (exception, stackTrace) async {
-      // runZonedGuarded doesn't crash the App.
-      final mechanism = Mechanism(type: 'runZonedGuarded', handled: true);
-      final throwableMechanism = ThrowableMechanism(mechanism, exception);
+    runZonedGuarded(
+      () async {
+        await _runner();
+      },
+      (exception, stackTrace) async {
+        // runZonedGuarded doesn't crash the App.
+        final mechanism = Mechanism(type: 'runZonedGuarded', handled: true);
+        final throwableMechanism = ThrowableMechanism(mechanism, exception);
 
-      final event = SentryEvent(
-        throwable: throwableMechanism,
-        level: SentryLevel.fatal,
-      );
+        final event = SentryEvent(
+          throwable: throwableMechanism,
+          level: SentryLevel.fatal,
+        );
 
-      await hub.captureEvent(event, stackTrace: stackTrace);
-    });
+        await hub.captureEvent(event, stackTrace: stackTrace);
+      },
+      zoneSpecification: ZoneSpecification(
+        print: (self, parent, zone, line) {
+          if (!hub.isEnabled) {
+            // early bail out, in order to better guard against the recursion
+            // as decribed below.
+            parent.print(zone, line);
+            return;
+          }
+
+          if (_isPrinting) {
+            // We somehow landed in a recursion.
+            // This happens for example if:
+            // hub.addBreadcrumb() called print() itself.
+            // This happens for example if hub.isEnabled == false and
+            // options.logger == dartLogger
+            //
+            // Anyway, in order to not cause a stack overflow due to recursion
+            // we drop any print() call while adding a breadcrumb.
+            return;
+          }
+          _isPrinting = true;
+          try {
+            hub.addBreadcrumb(Breadcrumb(
+              message: line,
+            ));
+            parent.print(zone, line);
+          } finally {
+            _isPrinting = false;
+          }
+        },
+      ),
+    );
 
     options.sdk.addIntegration('runZonedGuardedIntegration');
   }