+++ dart/test/default_integrations_test.dart
@@ -108,9 +108,65 @@ void main() {
     expect(true, throwableMechanism.mechanism.handled);
     expect(throwable, throwableMechanism.throwable);
   });
+
+  test('Run zoned guarded logs calls to print as breadcrumb', () async {
+    Future<void> callback() async {
+      print('foo bar');
+    }
+
+    final integration = RunZonedGuardedIntegration(callback);
+
+    await integration(fixture.hub, fixture.options);
+
+    expect(fixture.hub.addBreadcrumbCalls.length, 1);
+    final breadcrumb = fixture.hub.addBreadcrumbCalls.first.crumb;
+    expect(breadcrumb.message, 'foo bar');
+  });
+
+  test('Run zoned guarded: No addBreadcrumb calls for disabled Hub', () async {
+    await fixture.hub.close();
+
+    Future<void> callback() async {
+      print('foo bar');
+    }
+
+    final integration = RunZonedGuardedIntegration(callback);
+
+    await integration(fixture.hub, fixture.options);
+
+    expect(fixture.hub.addBreadcrumbCalls.length, 0);
+  });
+
+  test('Run zoned guarded: No recursion for print() calls', () async {
+    final options = SentryOptions(dsn: fakeDsn);
+    final hub = PrintRecursionMockHub();
+
+    Future<void> callback() async {
+      print('foo bar');
+    }
+
+    final integration = RunZonedGuardedIntegration(callback);
+
+    await integration(hub, options);
+
+    expect(hub.addBreadcrumbCalls.length, 1);
+    final breadcrumb = hub.addBreadcrumbCalls.first.crumb;
+    expect(breadcrumb.message, 'foo bar');
+  });
 }
 
 class Fixture {
   final hub = MockHub();
   final options = SentryOptions(dsn: fakeDsn);
+}
+
+class PrintRecursionMockHub extends MockHub {
+  @override
+  bool get isEnabled => true;
+
+  @override
+  void addBreadcrumb(Breadcrumb crumb, {dynamic hint}) {
+    print('recursion');
+    super.addBreadcrumb(crumb, hint: hint);
+  }
 }