+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix: Use name from pubspec.yaml for release if package id is not available (#411)
 * Feat: `SentryHttpClient` tracks the duration which a request takes and logs failed requests (#414)
 * Fix: Trim `\u0000` from Windows package info (#420)
+* Feature: Log calls to `print()` as Breadcrumbs (#439)
 
 # 5.0.0